+++ src/ska_mid_dish_manager/component_managers/dish_manager_cm.py
@@ -20,6 +20,7 @@ from ska_mid_dish_manager.models.dish_enums import (
     Band,
     BandInFocus,
     CapabilityStates,
+    Device,
     DishMode,
     DSOperatingMode,
     DSPowerState,
@@ -109,6 +110,12 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         self._state_update_lock = Lock()
         self._sub_communication_state_change_lock = Lock()
 
+        self._device_to_comm_attr_map = {
+            Device.DS: "dsConnectionState",
+            Device.SPF: "spfConnectionState",
+            Device.SPFRX: "spfrxConnectionState",
+        }
+
         # SPF has to go first
         self.sub_component_managers = {
             "SPF": SPFComponentManager(
@@ -126,7 +133,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 b5acapabilitystate=SPFCapabilityStates.UNAVAILABLE,
                 b5bcapabilitystate=SPFCapabilityStates.UNAVAILABLE,
                 communication_state_callback=partial(
-                    self._sub_communication_state_changed, "spfConnectionState"
+                    self._sub_communication_state_changed, Device.SPF
                 ),
                 component_state_callback=self._component_state_changed,
                 quality_state_callback=self._quality_state_callback,
@@ -152,7 +159,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 actstaticoffsetvaluexel=None,
                 actstaticoffsetvalueel=None,
                 communication_state_callback=partial(
-                    self._sub_communication_state_changed, "dsConnectionState"
+                    self._sub_communication_state_changed, Device.DS
                 ),
                 component_state_callback=self._component_state_changed,
                 quality_state_callback=self._quality_state_callback,
@@ -175,7 +182,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
                 b5acapabilitystate=SPFRxCapabilityStates.UNKNOWN,
                 b5bcapabilitystate=SPFRxCapabilityStates.UNKNOWN,
                 communication_state_callback=partial(
-                    self._sub_communication_state_changed, "spfrxConnectionState"
+                    self._sub_communication_state_changed, Device.SPFRX
                 ),
                 component_state_callback=self._component_state_changed,
                 quality_state_callback=self._quality_state_callback,