+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 ## unreleased
 *************
+- Updated `buildState` attribute to include version information of dish manager and subservient devices
 - Upgraded ska-mid-dish-simulators chart to v4.0.1
 - Upgraded ska-mid-ds-manager version to v2.0.0
 - Added actStaticOffsetValueXel and actStaticOffsetValueEl attributes